correction d'un problème avec l'horloge lorsque le tempo changeait

+++ TP_AOC_FX/src/TP_AOC/v1/Engine/Horloge/Horloge.java
@@ -9,7 +9,7 @@ public interface Horloge {
     /**
      * permet de changer la fréquence de répétition
      */
-    public void changerFrequence(int frequence);
+    public void changerFrequence(double frequence);
     /**
      * permet de démarrer à la fréquence donner préalablement
      */
